requirements.txt
pycryptodomex==3.14.1
jsonpickle==2.1.0
ecdsa==0.17.0
python-dotenv==0.20.0
python-daemon==2.3.0
pypiwin32
Flask==2.0.2